[lib] Support a profile setting in the annocheck config block
Allow the configuration file to provide a libannocheck profile that
rpminspect should use at runtime.  The idea here is that the vendor
data package can specify this in its configuration file, or users can
specify one in their rpminspect.yaml file.  The default mode of
operation is to not specify a profile.

Signed-off-by: David Cantrell <dcantrell@redhat.com>