[bld] Add 'Recommends: system-rpm-config' to the spec file template
librpminspect uses librpm to preprocess and parse spec files, but to
do that correctly requires the host system have spec file macro
definitions in use.  That's going to vary by RPM environment, but in
the Fedora/CentOS/RHEL sphere, we can put a weak dependency on
system-rpm-config and pull in macro definitions.  This is similar to
what the rpm-build package does.

Fixes: #1526

Signed-off-by: Dave Cantrell <dcantrell@redhat.com>